now its fixed

+++ src/alias.h
@@ -155,7 +155,7 @@ bool IsAliasOp(int op);
 
 int64 GetNameTxPosHeight(const CAliasIndex& txPos);
 int64 GetNameTxPosHeight(const CDiskTxPos& txPos);
-int64 GetNameTxPosHeight2(const CDiskTxPos& txPos, int nHeight);
+int64 GetNameTxPosHeight2(const CDiskTxPos& txPos, int64 nHeight);
 bool GetTxOfAlias(CAliasDB& dbName, const std::vector<unsigned char> &vchName, CTransaction& tx);
 int IndexOfNameOutput(const CTransaction& tx);
 bool GetValueOfNameTxHash(const uint256& txHash, std::vector<unsigned char>& vchValue, uint256& hash, int& nHeight);
